Folder restructuring.
