+++ ecal/core/src/registration/shm/ecal_registration_receiver_shm.h
